+++ packages/featureserver/src/query.js
@@ -16,6 +16,8 @@ function query (data, params = {}) {
   // TODO clean up this series of if statements
   const filtersApplied = data.filtersApplied || {}
   const options = _.cloneDeep(params)
+  const hasIdField = _.has(data, 'metadata.idField')
+
   if (filtersApplied.projection) delete options.outSR
   if (filtersApplied.geometry) delete options.geometry
   if (filtersApplied.where || options.where === '1=1') delete options.where
@@ -30,13 +32,23 @@ function query (data, params = {}) {
   if (options.f !== 'geojson') options.toEsri = true
   const queriedData = filtersApplied.all ? data : Winnow.query(data, options)
 
+  // ArcGIS client warnings
+  if (options.toEsri) {
+    if (!hasIdField) console.warn(`The requested provider has no "idField" assignment. This can cause errors in ArcGIS clients`)
+    else if (data.metadata.idField.toLowerCase() === 'objectid' && data.metadata.idField !== 'OBJECTID') {
+      console.warn(`The requested provider's "idField" is a mixed-case version of "OBJECTID". This can cause errors in ArcGIS clients`)
+    } else if (queriedData.features.some(feature => { return !Number.isInteger(feature.attributes.OBJECTID) || feature.attributes.isArrayOBJECTID > 2147483647 })) {
+      console.warn(`OBJECTIDs created from provider's "idField" are not integers from 0 to 2147483647`)
+    }
+  }
+
   if (params.f === 'geojson') return { type: 'FeatureCollection', features: queriedData.features }
   else return geoservicesPostQuery(data, queriedData, params)
 }
 
 function geoservicesPostQuery (data, queriedData, params) {
   // options.objectIds works alongside returnCountOnly but not statistics
-  const oidField = (data.metadata && data.metadata.idField) || 'OBJECTID'
+  const oidField = 'OBJECTID'
   if (params.objectIds && !params.outStatistics) {
     let oids
 
@@ -69,7 +81,7 @@ function geoservicesPostQuery (data, queriedData, params) {
 }
 
 function idsOnly (data, options = {}) {
-  const oidField = options.idField || 'OBJECTID'
+  const oidField = 'OBJECTID'
   return data.features.reduce(
     (resp, f) => {
       resp.objectIds.push(f.attributes[oidField])