+++ packages/featureserver/src/templates.js
@@ -54,7 +54,7 @@ function renderLayer (featureCollection = {}, options = {}) {
   if (json.timeInfo) json.timeInfo = metadata.timeInfo
   if (json.maxRecordCount) json.maxRecordCount = metadata.maxRecordCount || 2000
   if (json.displayField) json.displayField = metadata.displayField || json.fields[0].name
-  if (json.objectIdField) json.objectIdField = metadata.idField || 'OBJECTID'
+  if (json.objectIdField) json.objectIdField = 'OBJECTID'
   if (capabilities.quantization) json.supportsCoordinatesQuantization = true
   return json
 }