+++ packages/featureserver/test/template.json.js
@@ -110,7 +110,7 @@ describe('Template content', () => {
         'globalIdField': Joi.string().valid(''),
         'types': Joi.array().min(0),
         'templates': Joi.array().min(0),
-        'hasStaticData': Joi.boolean().valid(true),
+        'hasStaticData': Joi.boolean().valid(false),
         'timeInfo': Joi.object().keys({})
       })
       Joi.validate(layerJson, schema, {presence: 'required'}).should.have.property('error', null)