+++ packages/featureserver/CHANGELOG.md
@@ -2,6 +2,15 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## Unreleased
+### Added
+* provider attributes referenced by metadata `idField` are maintained as separate field in addition to OBJECTID when creating ESRI json
+* warnings when a provider's `idField` is not set or references non- or out-of-range integer values
+* warning when a provider's `idField` is a mixed-case version of 'OBJECTID'
+
+### Fixed
+* changed default value of `hasStaticData` to `false`
+
 ## [2.10.2] - 04-10-2018
 ### Fixed 
 * set Content-Type: application/javascript when response is wrapped in callback (JSONP)